+++ src/myinf/configs.nim
@@ -1,0 +1,6 @@
+from std/os import getEnv
+
+const
+  dbHost* = getEnv "DB_HOST"
+  dbUser* = getEnv "DB_USER"
+  dbPass* = getEnv "DB_PASS"