+++ src/myinf/db/setup.nim
@@ -1,0 +1,5 @@
+from myinf/db import DbConn, createTables
+from myinf/db/models/user import newUser
+
+proc setup*(conn: DbConn) =
+  conn.createTables(newUser())