+++ src/myinf/routes/helloWorld.nim
@@ -1,0 +1,4 @@
+import pkg/prologue
+
+proc r_helloWorld*(ctx: Context) {.async.} =
+  resp "<h1>Hello, Prologue!</h1>"