+++ changelog.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Version 0.2.0 (Jul 10 2022)
+
+- DB
+  - Added db connection
+  - Added user model
+  - Added setup
+- added Hello world route
+- Added routes file
+- Added env loader in `configs.nim`
+
+---
+
 ## Version 0.1.0 (Jul 9 2022)
 
 - Init